update ban system and f5 administration

+++ src/components/ban/server/removePlayerBan.lua
@@ -9,10 +9,20 @@
 --]]
 ---@author VibR1cY
 
-_FlashServer_Bans.removeBan = function(identifier)
+_FlashServer_Bans.removePlayerBan = function(identifier, callback)
+    if (not (_FlashServer_Bans.exist(identifier))) then
+        return callback(false)
+    end
     _FlashServer_Database.execute("DELETE FROM flash_bans WHERE identifier = @identifier", {
         ["identifier"] = identifier
     }, function()
         _FlashServer_Bans.remove(identifier)
+        callback(true)
     end)
 end
+
+RegisterCommand("unban", function(source, args)
+    _FlashServer_Bans.removePlayerBan(args[1], function(sucess)
+        print(sucess)
+    end)
+end)
